all function and popups working just missing the like function and the save to saving fucntion

+++ src/pages/index.js
@@ -33,31 +33,37 @@ const initialCards = [
     name: "Yosemite Valley",
     description: "Yosemite Valley",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/yosemite.jpg",
+    _id: "card1",
   },
   {
     name: "Lake Louise",
     description: "Lake Louise",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/lake-louise.jpg",
+    _id: "card12",
   },
   {
     name: "Bald Mountains",
     description: "Bald Mountains",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/bald-mountains.jpg",
+    _id: "card123",
   },
   {
     name: "Latemar",
     description: "Latemar",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/latemar.jpg",
+    _id: "card4",
   },
   {
     name: "Vanoise National Park",
     description: "Vanoise National Park",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/vanoise.jpg",
+    _id: "card5",
   },
   {
     name: "Lago di Braies",
     description: "Lago di Braies",
     link: "https://practicum-content.s3.us-west-1.amazonaws.com/software-engineer/around-project/lago.jpg",
+    _id: "card6",
   },
 ];
 //
@@ -100,6 +106,15 @@ const deleteCardPopup = new PopupWithConfirmation({
   popupSelector: "#delete__card-modal",
   handleFormSubmit: (cardId, cardElement) => {
     console.log(`Deleting card with cardId=${cardId}`);
+    // Check if card ID is from a hardcoded card
+    if (cardId.startsWith("card")) {
+      console.warn("Cannot delete hardcoded initial card");
+      cardElement.remove();
+      deleteCardPopup.close();
+      return;
+    }
+
+    // Proceed to delete card from server
     api
       .deleteCard(cardId)
       .then(() => {
@@ -129,6 +144,11 @@ const createCard = (cardData) => {
 };
 
 // Card Section initialization
+
+// Fetch and Render Cards from Server
+
+// Fetch and Render Cards from Server
+
 const cardSection = new Section(
   {
     items: initialCards,
@@ -141,14 +161,16 @@ const cardSection = new Section(
   ".cards__list"
 );
 cardSection.renderItems();
+
 // Form Handlers
 
 const handleProfileFormSubmit = (data) => {
   api
     .editProfile(data.name, data.job)
     .then((res) => {
-      console.log("Profile Response:", res); // Log the response to see the structure
-      userInfo.setUserInfo({ name: res.name, job: res.about });
+      console.log("Profile Response:", res);
+      profileTitle.textContent = data.name;
+      profileDescription.textContent = data.job;
       editProfilePopup.close();
     })
     .catch((err) => {
@@ -227,5 +249,4 @@ profileEditButton.addEventListener("click", () => {
 
 profileAvatarEditButton.addEventListener("click", () => {
   editAvatarPopup.open();
-  avatarEditFormValidator.reset();
 });
